# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ExtractIQ MCQ Platform API
  description: |
    REST API for the ExtractIQ MCQ Platform — an AI-powered Multiple Choice Question
    extraction and management system. Supports document upload, AI-driven MCQ generation,
    human review workflows, and multi-format export.
  version: 1.0.0
  contact:
    name: ExtractIQ Team
servers:
  - url: /api/v1
    description: API v1
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        statusCode:
          type: integer
    Success:
      type: object
      properties:
        success:
          type: boolean
...
    ChangePasswordRequest:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
          minLength: 12
    AcceptInvitationRequest:
      type: object
      required: [token, name, password]
      properties:
        token:
          type: string
        name:
          type: string
        password:
          type: string
          minLength: 12
    TokenPair:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
    UserProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        name:
          type: string
        role:
          type: string
        workspaceId:
          type: string
          format: uuid
...
    PresignUploadRequest:
      type: object
      required: [filename, contentType, fileSize, projectId]
      properties:
        filename:
          type: string
        contentType:
          type: string
        fileSize:
          type: integer
        projectId:
          type: string
          format: uuid
    CompleteUploadRequest:
      type: object
      required: [uploadId, s3Key]
      properties:
        uploadId:
          type: string
        s3Key:
          type: string
        checksumSha256:
          type: string
...
    DocumentResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        filename:
          type: string
        contentType:
          type: string
        fileSize:
          type: integer
        status:
          type: string
        projectId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
...
  responses:
    Unauthorized:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
paths:
...
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      operationId: refreshToken
      responses:
        '200':
          description: New token pair
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/change-password:
    post:
      tags: [Auth]
      summary: Change current user password
      operationId: changePassword
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200':
          description: Password changed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/invitations/{token}:
    get:
      tags: [Auth]
      summary: Get invitation metadata by token
      operationId: getInvitation
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Invitation metadata
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      email:
                        type: string
                        format: email
                      role:
                        type: string
                      workspaceId:
                        type: string
                        format: uuid
                      workspaceName:
                        type: string
                      invitedName:
                        type: string
                      expiresAt:
                        type: string
                        format: date-time
                      accepted:
                        type: boolean
                      expired:
                        type: boolean
                      status:
                        type: string
                        enum: [pending, accepted, expired]
        '404':
          $ref: '#/components/responses/NotFound'
  /auth/accept-invitation:
    post:
      tags: [Auth]
      summary: Accept an invitation and create an authenticated session
      operationId: acceptInvitation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AcceptInvitationRequest'
      responses:
        '200':
          description: Invitation accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      user:
                        $ref: '#/components/schemas/UserProfile'
                      token:
                        type: string
        '404':
          $ref: '#/components/responses/NotFound'
...
  /documents/presign:
    post:
      tags: [Documents]
      summary: Get presigned upload URL
      operationId: presignUpload
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PresignUploadRequest'
      responses:
        '201':
          description: Presigned URL and document upload metadata
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      uploadUrl:
                        type: string
                        format: uri
                      documentId:
                        type: string
                        format: uuid
                      s3Key:
                        type: string
                      expiresIn:
                        type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
  /documents/complete:
    post:
      tags: [Documents]
      summary: Complete upload and create document record
      operationId: completeUpload
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompleteUploadRequest'
      responses:
        '200':
          description: Document upload completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/DocumentResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...